large refactor, somewhat breaking:  - implement more structure to transactions   - allow new sessions to be blocked from initalization   - allow tx's to check for R/W flags   - track outstanding global 'admin operations'  1st pass implementation of locks & counters
 cleanup of KvTx logic

+++ src/main/groovy/net/iowntheinter/kvdn/storage/kv/impl/KvTx.groovy
@@ -9,6 +9,7 @@ import io.vertx.core.logging.Logger
 import io.vertx.core.shareddata.AsyncMap
 import io.vertx.core.shareddata.SharedData
 import io.vertx.core.logging.LoggerFactory
+import net.iowntheinter.kvdn.kvdnTX
 import net.iowntheinter.kvdn.storage.kv.TXKV
 import net.iowntheinter.kvdn.storage.kv.local.shimAsyncMap
 
@@ -18,16 +19,21 @@ import java.security.MessageDigest
  * Created by grant on 11/19/15.
  */
 
-class KvTx implements TXKV {
+class KvTx extends kvdnTX implements TXKV {
     SharedData sd
     Logger logger
     EventBus eb
     String strAddr
+    UUID txid
     def  keyprov
     def Vertx vertx
     def D = new _data_impl()
     def session
-    class _data_impl {
+    enum txtype {
+        MODE_WRITE,
+        MODE_READ
+    }
+    private class _data_impl {
         SharedData sd;
         Vertx vertx
         String name
@@ -46,11 +52,12 @@ class KvTx implements TXKV {
         }
     }
 
-    def KvTx(String sa, kvdnSession session, Vertx vertx) {
+    def KvTx(String sa, UUID txid, kvdnSession session, Vertx vertx) {
         // keys = new ORSet()
         this.vertx = vertx
         this.keyprov = session.keyprov
-        this.session = session
+        this.session = session as kvdnSession
+        this.txid = txid
         strAddr = sa
         logger = new LoggerFactory().getLogger("KvTx:" + strAddr)
         sd = vertx.sharedData() as SharedData
@@ -59,16 +66,23 @@ class KvTx implements TXKV {
 
     @Override
     Object bailTx(Object context) {
+        logger.error("KVTX error: ${this.session} " + context as Map)
+        (this.session as kvdnSession).finishTx(this)
         return null
     }
-
+    Set getFlags(){
+        return session.txflags
+    }
+    boolean checkFlags(txtype){
+        return (!session.txflags.contains(txtype))
+    }
     @Override
     void snapshot() {}
 
     @Override
     void submit(content, cb) {
         D.getMap(this, { res ->
-            if (res.succeeded()) {
+            if (res.succeeded() && checkFlags(txtype.MODE_WRITE)) {
                 def AsyncMap map = res.result();
                 def String key = MessageDigest.getInstance("MD5").digest(Buffer.buffer(content.toString()).getBytes()).encodeHex().toString()
 
@@ -78,19 +92,18 @@ class KvTx implements TXKV {
                         keyprov.addKey(strAddr, key, {
                             logger.trace("set:${strAddr}:${key}");
                             eb.publish("_KVDN_+${strAddr}", new JsonObject().put('key',key))
+                            (this.session as kvdnSession).finishTx(this)
                             cb([result: resPut.result().toString(), key: key, error: null])
                         })
 
                     } else {
-                        logger.error("an error occured in a submit() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
-
+                        bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                         cb([result: null, error: resPut.cause()])
                     }
                 })
             } else {
-                logger.error("an error occured in a submit() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
-
-                cb([result: null, error: res.cause()])
+                bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
+                cb([result: null, error: (res.cause() ?: getFlags())])
             }
         })
     }
@@ -98,80 +111,102 @@ class KvTx implements TXKV {
     @Override
     void set(String key, content, cb) {
         D.getMap(this, { res ->
-            if (res.succeeded()) {
+            if (res.succeeded() && checkFlags(txtype.MODE_WRITE)) {
                 def AsyncMap map = res.result();
                 map.put(key, content, { resPut ->
                     if (resPut.succeeded()) {
                         keyprov.addKey(strAddr, key, {
                             logger.trace("set:${strAddr}:${key}");
                             eb.publish("_KVDN_+${strAddr}", new JsonObject().put('key',key))
+                            (this.session as kvdnSession).finishTx(this)
                             cb([result: resPut.result().toString(), key: key, error: null])
                         })
                     } else {
-                        logger.error("an error occured in a set() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
+                        bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                         cb([result: null, error: resPut.cause()])
                     }
                 })
             } else {
-                logger.error("an error occured in a set() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
+                bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                 cb([result: null, error: res.cause()])
             }
         })
     }
 
-    @Override
-    void getKeys(cb) {
-        keyprov.getKeys(strAddr, { Map asyncResult ->
 
-            cb(asyncResult)
-        })
-    }
 
     @Override
     void get(String key, cb) {
         D.getMap(this, { res ->
-            if (res.succeeded()) {
+            if (res.succeeded() && checkFlags(txtype.MODE_READ)) {
                 def AsyncMap map = res.result();
                 map.get(key, { resGet ->
                     if (resGet.succeeded()) {
-                        logger.trace("get:${strAddr}:${key}");
+                        logger.trace("get:${strAddr}:${key}")
+                        (this.session as kvdnSession).finishTx(this)
                         cb([result: resGet.result().toString(), error: null])
                     } else {
-                        logger.error("an error occured in a get() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
+                        bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                         cb([result: null, error: resGet.cause()])
                     }
                 })
             } else {
+                bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                 cb([result: null, error: res.cause()])
             }
         })
     }
-
     @Override
     void del(String key, cb) {
         D.getMap(this, { res ->
-            if (res.succeeded()) {
+            if (res.succeeded() && checkFlags(txtype.MODE_WRITE)) {
                 def AsyncMap map = res.result();
                 map.remove(key, { resDel ->
                     if (resDel.succeeded()) {
                         keyprov.deleteKey(strAddr, key, {
                             logger.trace("set:${strAddr}:${key}");
                             eb.publish("_KVDN_-${strAddr}", new JsonObject().put('key'))
+                            (this.session as kvdnSession).finishTx(this)
                             cb([result: resDel.result().toString(), key: key, error: null])
                         })
-                        cb([result: resDel.result().toString(), error: null])
-
                     } else {
-                        logger.error("an error occured in a del() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
-
+                        bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                         cb([result: null, error: resDel.cause()])
                     }
                 })
             } else {
-                logger.error("an error occured in a del() transaction on key: ${key} straddr: ${strAddr} session: ${this.session} ")
-
+                bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
                 cb([result: null, error: res.cause()])
             }
         })
     }
+    @Override
+    void getKeys(cb) {
+        keyprov.getKeys(strAddr, { Map asyncResult ->
+            (this.session as kvdnSession).finishTx(this)
+            cb(asyncResult)
+        })
+    }
+    @Override
+    void size(cb) {
+        D.getMap(this, { res ->
+            if (res.succeeded() && checkFlags(txtype.MODE_READ)) {
+                def AsyncMap map = res.result();
+                map.size( { resGet ->
+                    if (resGet.succeeded()) {
+                        logger.trace("size:${strAddr}")
+                        (this.session as kvdnSession).finishTx(this)
+                        cb([result: resGet.result().toString(), error: null])
+                    } else {
+                        bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
+                        cb([result: null, error: resGet.cause()])
+                    }
+                })
+            } else {
+                bailTx([txid:this.txid,straddr:strAddr,session:this.session,flags:session.txflags])
+                cb([result: null, error: res.cause()])
+            }
+        })
+    }
+
 }
